+++ R/custom_dims_mets.R
@@ -67,7 +67,7 @@ allowed_metric_dim <- function(type = c("METRIC", "DIMENSION"),
 #'
 #' @return Custom Metric or Dimension meta data
 #' @importFrom googleAuthR gar_api_generator
-#' @family managementAPI functions
+#' @family custom variable functions
 #' @export
 ga_custom_vars <- function(accountId,
                            webPropertyId,
@@ -95,8 +95,9 @@ ga_custom_vars <- function(accountId,
                                path_args = pa,
                                data_parse_function = function(x) x)
   
-  cus_var()
+  res <- cus_var()
   
+  as.customDimension(res)
 }
 
 #' List Custom Dimensions or Metrics
@@ -107,8 +108,24 @@ ga_custom_vars <- function(accountId,
 #'
 #' @return Custom Metric or Dimension List
 #' @importFrom googleAuthR gar_api_generator
-#' @family managementAPI functions
+#' @family custom variable functions
 #' @export
+#' 
+#' @details 
+#' 
+#' This function lists all the existing custom dimensions or metrics for the web property. 
+#' 
+#' @examples 
+#' 
+#' \dontrun{
+#' library(googleAnalyticsR)
+#' ga_auth()
+#' 
+#' ga_custom_vars_list(54019251, webPropertyId = "UA-54019251-4", type = "customDimensions")
+#' 
+#' ga_custom_vars_list(54019251, webPropertyId = "UA-54019251-4", type = "customMetrics")
+#' 
+#' }
 ga_custom_vars_list <- function(accountId,
                                 webPropertyId,
                                 type = c("customDimensions","customMetrics")){
@@ -160,4 +177,205 @@ parse_ga_custom_vars_list <- function(x){
   attr(o, "nextLink") <- x$nextLink
   o
   
+}
+
+#' Create a custom dimension
+#' 
+#' Create a dimension using \link{customDimension} then pass it into this function to create it in your web property.
+#' 
+#' @inheritParams customDimension
+#' 
+#' @seealso \href{https://support.google.com/analytics/answer/2709828}{Custom dimensions support article}
+#' @export
+#' @importFrom googleAuthR gar_api_generator
+#' @family custom variable functions
+#' 
+#' 
+#' @examples 
+#' 
+#' \dontrun{
+#' library(googleAnalyticsR)
+#' ga_auth()
+#' 
+#' # create custom var
+#' ga_custom_vars_create("my_custom_dim",
+#'                       index = 15,
+#'                       accountId = 54019251,
+#'                       webPropertyId = "UA-54019251-4",
+#'                       scope = "HIT",
+#'                       active = FALSE)
+#' 
+#' # view custom dimension in list
+#' ga_custom_vars_list(54019251, webPropertyId = "UA-54019251-4", type = "customDimensions")
+#' 
+#' }
+ga_custom_vars_create <- function(name, 
+                                  index,
+                                  accountId,
+                                  webPropertyId,
+                                  active,
+                                  scope = c("HIT","SESSION","USER","PRODUCT")){
+  scope <- match.arg(scope)
+  accountId <- as.character(accountId)
+  
+  cd <- customDimension(name = name,
+                        index = index,
+                        accountId = accountId,
+                        webPropertyId = webPropertyId,
+                        scope = scope,
+                        active = active)
+
+  api <- gar_api_generator("https://www.googleapis.com/analytics/v3/management",
+                           "POST",
+                           path_args = list(
+                             accounts = accountId,
+                             webproperties = webPropertyId,
+                             customDimensions = ""
+                           ),
+                           data_parse_function = function(x) x)
+  api(the_body = cd)
+  
+  
+}
+
+#' Modify a custom dimension
+#' 
+#' Modify existing custom dimensions
+#' 
+#' @inheritParams customDimension
+#' @param ignoreCustomDataSourceLinks Force the update and ignore any warnings related to the custom dimension being linked to a custom data source / data set.
+#' 
+#' @seealso \href{https://support.google.com/analytics/answer/2709828}{Custom dimensions support article}
+#' @export
+#' @import assertthat
+#' @importFrom googleAuthR gar_api_generator
+#' @family custom variable functions
+#' 
+#' 
+#' @examples 
+#' 
+#' \dontrun{
+#' library(googleAnalyticsR)
+#' ga_auth()
+#' 
+#' # create custom var
+#' ga_custom_vars_create("my_custom_dim",
+#'                       index = 15,
+#'                       accountId = 54019251,
+#'                       webPropertyId = "UA-54019251-4",
+#'                       scope = "HIT",
+#'                       active = FALSE)
+#' 
+#' # view custom dimension in list
+#' ga_custom_vars_list(54019251, webPropertyId = "UA-54019251-4", type = "customDimensions")
+#' 
+#' # change a custom dimension
+#' ga_custom_vars_patch("ga:dimension7",
+#'                      accountId = 54019251,
+#'                      webPropertyId = "UA-54019251-4",
+#'                      name = "my_custom_dim2",
+#'                      scope = "SESSION",
+#'                      active = TRUE)
+#'                      
+#' # view custom dimensions again to see change
+#' ga_custom_vars_list(54019251, webPropertyId = "UA-54019251-4", type = "customDimensions")
+#' 
+#' }
+ga_custom_vars_patch <- function(id,
+                                 accountId,
+                                 webPropertyId,
+                                 name = NULL, 
+                                 active = NULL,
+                                 scope = NULL,
+                                 ignoreCustomDataSourceLinks = FALSE){
+  
+  accountId <- as.character(accountId)
+  assert_that(is.string(id))
+  
+  if(!is.null(scope)){
+    assert_that(is.string(scope),
+                scope %in% c("HIT","SESSION","USER","PRODUCT"))
+  }
+  
+  
+  cd <- customDimension(name = name,
+                        scope = scope,
+                        active = active,
+                        accountId = accountId,
+                        webPropertyId = webPropertyId)
+  
+  cd <- rmNullObs(cd)
+  
+  api <- gar_api_generator("https://www.googleapis.com/analytics/v3/management/",
+                           "PATCH",
+                           path_args = list(accounts = accountId,
+                                            webproperties = webPropertyId,
+                                            customDimensions = id),
+                           pars_args = list(
+                             ignoreCustomDataSourceLinks = ignoreCustomDataSourceLinks
+                             ),
+                           data_parse_function = function(x) x)
+  res <- api(the_body = cd)
+  as.customDimension(res)
+}
+
+
+#' A customDimension object
+#' 
+#' Pass this into \link{ga_custom_vars_create} or \link{ga_custom_var_update}
+#' 
+#' @param active \code{TRUE} or \code{FALSE} if custom dimension is active or not
+#' @param index Index of custom dimension - integer between 1 and 20 (200 for GA360)
+#' @param name Name of custom dimension
+#' @param scope Scope of custom dimension - one of \code{"HIT","SESSION","USER","PRODUCT"}
+#' @param accountId AccountId of the custom dimension
+#' @param webPropertyId WebPropertyId of the custom dimension
+#' @param id The id of the custom dimension
+#' 
+#' @details 
+#'  When creating a new custom dimension, supply all the arguments except id.  
+#'  
+#'  If you are updating an existing custom dimension, only supply the arguments you wish to change
+#' 
+#' @family custom variable functions
+#' @import assertthat
+customDimension <- function(name = NULL, 
+                            index = NULL,
+                            scope = NULL, 
+                            active = NULL,
+                            id = NULL,
+                            accountId = NULL,
+                            webPropertyId = NULL){
+  
+  assert_that_ifnn(name, is.string)
+  assert_that_ifnn(scope, is.string)
+  assert_that_ifnn(active, is.flag)
+  assert_that_ifnn(id, is.string)
+  
+  if(!is.null(index)){
+    index <- as.numeric(index)
+    assert_that(is.scalar(index),
+                index <= 200,
+                index > 0)
+  }
+  
+  structure(
+    list(id = id,
+         name = name,
+         index = index,
+         scope = scope,
+         active = active,
+         accountId = accountId,
+         webPropertyId = webPropertyId),
+    class = "customDimension_ga"
+  )
+}
+
+is.customDimension <- function(x){
+  inherits(x, "customDimension_ga")
+}
+
+as.customDimension <- function(x){
+  class(x) <- c("customDimension_ga", class(x))
+  x
 }